requirements.txt
# Astro Runtime includes the following pre-installed providers packages: 
#https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

pandas
requests
dbt-snowflake
apache-airflow-providers-snowflake
airflow-provider-fivetran-async
datetime
boto3
apache-airflow-providers-amazon>=9.0.0
streamlit
plotly
cryptography

